+++ v2kit/__init__.py
@@ -2,7 +2,7 @@
 """v2kit modules."""
 from .params import V2KIT_VERSION, Protocol
 from .errors import V2kitError, V2kitValidationError, V2kitParseError
-from .functions import is_vmess, is_vless, is_trojan, is_shadowsocks, is_socks
+from .functions import is_vmess, is_vless, is_trojan, is_shadowsocks, is_socks, is_http
 from .functions import relabel
 from .functions import encode_subscription, decode_subscription
 from .parsers import parse
@@ -12,6 +12,7 @@ from .models import (
     TrojanConfig,
     ShadowsocksConfig,
     SocksConfig,
+    HttpConfig,
 )
 __version__ = V2KIT_VERSION
 
@@ -24,6 +25,7 @@ __all__ = [
     "is_trojan",
     "is_shadowsocks",
     "is_socks",
+    "is_http",
     "relabel",
     "encode_subscription",
     "decode_subscription",
@@ -33,4 +35,5 @@ __all__ = [
     "VLESSConfig",
     "TrojanConfig",
     "ShadowsocksConfig",
-    "SocksConfig"]
+    "SocksConfig",
+    "HttpConfig"]