+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `HttpConfig` class
+### Changed
+- `README.md` updated
+- Test system modified
 ## [0.3] - 2026-06-16
 ### Added
 - Logo